fix: move things around to avoid cyclic imports
