+++ sqlmesh/core/_typing.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import typing as t
-from datetime import datetime
 
 from sqlglot import exp
 
@@ -10,5 +9,3 @@ if t.TYPE_CHECKING:
     SchemaName = t.Union[str, exp.Table]
     SessionProperties = t.Dict[str, t.Union[exp.Expression, str, int, float, bool]]
     CustomMaterializationProperties = t.Dict[str, t.Union[exp.Expression, str, int, float, bool]]
-    Interval = t.Tuple[datetime, datetime]
-    Batch = t.List[Interval]
